benchmark/requirements.txt
requests>=2.31.0
pandas>=2.2.0
openpyxl>=3.1.2
tqdm>=4.66.0
datasets>=2.18.0
huggingface_hub>=0.21.0
gitpython>=3.1.40
langdetect>=1.0.9
anthropic>=0.40.0
